rename `z-shell` crate -> `shell`
