+++ crates/shell/src/encoding.rs
