+++ crates/shell/src/error.rs
