+++ crates/shell/src/options.rs
