+++ crates/shell/src/shell.rs
